Add GActionMuxer
This is a new implementation of an action muxer with the same interface as the
one in the gtk tree.  Its implementation is considerably simpler, mostly
because it doesn't need to implement GActionObservable.

In addition to muxing different action groups with the <prefix>.<action_name>
scheme, it has the notion of global, prefix-less actions.  Indicator-messages
needs those for the status and clear actions.