libmessaging-menu: use the (newly added) GTupleAction instead of GSimpleAction
GTupleAction is a bit simpler to handle when the action state contains a tuple
of things that are independently modified most the time.  It might be useful
for other indicators as well.

This implicitly also fixes the bug that libmessaging-menu did not preserve the
other values in the action state when updating count, time, or string.