Show menu separators between sections
Previously, gtk didn't support changing the menu model on an already-created
GtkMenu.  Since IndicatorObject doesn't allow changing the GtkMenu it exports
to the panel, IndicatorMessages created a menu with a single section, into
which the menu from the service was inserted (and removed when the service
died).  This led to seperators not being shown, because separators are only
shown between top-level sections.

Gtk now has gtk_menu_shell_bind_model, which allows us to get rid of top-level
wrapper section.