Change application to service d-bus protocol
Previously, the protocol was simply a menu model and an action group of the
currently active sources.  The service inserted the menu as a section into the
indicator menu.

This doesn't work anymore, because applications can (soon) expose individual
messages, and the messaging menu doesn't always display all of those at once.
This patch introduces a more specific d-bus API.

That API is still considered private: applications have to use
libmessaging-menu.