Notify the service when an app unrefs its MessagingMenuApp
Prior to this patch, the messaging menu only marked apps as "not running" when
they quit (i.e. disappeared from the bus).  This was okay, since most
applications only ever release the ref to their MessagingMenuApp when they
quit, or after calling _unregister explicitely (which removes them from the
menu entirely).

However, this is according to libmessagingmenu's documentation, and at least
indicator-telepathy relies on it.