application list: update root action unconditionally
Until now, the _update_root_action() was only called when the
draws_attention was set. That function now also sets the sensitive state
of some other actions, which might change independently of the attention
flag.

Simply call the function unconditionally. It's not doing a lot of work
anyway and it only gets called when something in the source or message
lists changes.